implemented channel search
