+++ app/controllers/api/channels_controller.rb
@@ -1,6 +1,6 @@
 class Api::ChannelsController < ApplicationController
     def index 
-        @channels = Channel.all
+        @channels = Channel.search(params[:query])
         render :index
     end
