+++ frontend/components/channels/channel_browser.jsx
@@ -4,10 +4,21 @@ import ChannelBrowserItem from "./channel_browser_item";
 export default class ChannelBrowser extends React.Component {
     constructor(props) {
         super(props);
+
+        this.handleInput = this.handleInput.bind(this);
+
+        this.state = ({
+            query: ""
+        })
     }
 
     componentDidMount() {
         this.props.fetchChannels();
+    }
+
+    handleInput(e) {
+        this.setState({query: e.target.value})
+        this.props.fetchChannels(e.target.value)
     }
 
     render()  {
@@ -15,19 +26,13 @@ export default class ChannelBrowser extends React.Component {
             <div className="user-dashboard-center-main">
                     <div className="user-dashboard-center-main-feed">
                         <div className="channel-browser-container">
-                        <input className="channel-browser-search-bar" type="text" placeholder='Search by channel name'/>
+                        <input className="channel-browser-search-bar" type="text" value={this.state.query} onChange={this.handleInput} placeholder='Search by channel name'/>
                             <span className="channels-browser-label">{this.props.channels.length} channels</span>
                              <ul>
                                 {this.props.channels.map((channel, idx) => (
                                     <ChannelBrowserItem key={`${channel.name}${idx}`} channel={channel}/>
                                 ))}
                             </ul>
-                        </div>
-                    </div>
-                    <div className="user-dashboard-center-main-message-composer">
-                        <div className="user-dashboard-message-container">
-                            <textarea type="text" className="message-input" placeholder="Write your message here..."/>
-                            <button className="send-message-button">Send</button>
                         </div>
                     </div>
             </div>
