+++ frontend/components/channels/channel_browser_container.js
@@ -1,13 +1,13 @@
 import { connect } from "react-redux";
 import ChannelBrowser from "./channel_browser";
-import { fetchChannels } from "../../actions/channel_actions";
+import { fetchChannels } from "../../actions/search_actions";
 
 const mapStateToProps = state => ({
-    channels: Object.values(state.entities.channels)
+    channels: Object.values(state.entities.searchResults.channels)
 })
 
 const mapDispatchToProps = dispatch => ({
-    fetchChannels: () => dispatch(fetchChannels())
+    fetchChannels: (query) => dispatch(fetchChannels(query))
 })
 
 const ChannelBrowserContainer = connect(mapStateToProps, mapDispatchToProps)(ChannelBrowser);
