+++ frontend/reducers/entities/search_results_reducer.js
@@ -1,4 +1,4 @@
-import { RECEIVE_SEARCHED_USERS, CLEAR_SEARCH_RESULTS } from "../../actions/search_actions";
+import { RECEIVE_SEARCHED_USERS, CLEAR_SEARCH_RESULTS, RECEIVE_CHANNELS } from "../../actions/search_actions";
 import { merge } from "lodash";
 
 const initialState = {
@@ -8,13 +8,18 @@ const initialState = {
 
 export const searchResultsReducer = (state = initialState, action) => {
     Object.freeze(state);
+    let nextState
     switch (action.type) {
         case RECEIVE_SEARCHED_USERS:
-            let nextState = merge({}, state);
+            nextState = merge({}, state);
             nextState.users = action.users;
             return nextState;
         case CLEAR_SEARCH_RESULTS:
             return initialState;
+        case RECEIVE_CHANNELS:
+            nextState = merge({}, state);
+            nextState.channels = action.channels
+            return nextState;
         default:
             return state;
     }
