--- frontend/util/channel_api_util.js
@@ -1,6 +1,0 @@
-export const fetchChannels = () => (
-    $.ajax({
-        method: "GET",
-        url: "/api/channels"
-    })
-);
